+++ src/screens/settings/PermissionScreen.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+import { Screen } from '../../components';
+
+function PermissionScreen(props) {
+    return (
+        <Screen>
+            <Text>Permissions</Text>
+        </Screen>
+    );
+}
+
+export default PermissionScreen;
